+++ addOns/authhelper/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - A toolbar button to the Authentication Diagnostics screenshots tab, to toggle screenshots between full size and scaled to fit the window.
+- Use the HTTP method for the verification configuration (on newer ZAP versions).
 
 ### Changed
 - Tag diagnostic HTTP messages with an internal ID, to make it easier to cross reference them.